Support NULL values through special string
Issue: When the data of a column is the empty string, the target currently writes it in as a NULL sql value. Violating not null invariant sometimes of the table schema.

Solution: Instead of treating empty string as NULL, we define an explicit NULL value str, use that in both writing to and reading from csv.

That way, empty strings remain empty strings